backend/services/autoplac_service.py
import requests
from flask import current_app

class AutoplacService:
    """Service for interacting with Autoplac.pl API"""
    
    def __init__(self):
        self.api_key = current_app.config.get('AUTOPLAC_API_KEY')
        self.api_url = current_app.config.get('AUTOPLAC_API_URL')
    
    def publish_car(self, car_data):
        """
        Publish a car listing to Autoplac.pl
        
        Args:
            car_data (dict): Car information to publish
            
        Returns:
            dict: Response from Autoplac API
        """
        # TODO: Implement actual Autoplac API integration
        # This is a placeholder for future implementation
        
        if not self.api_key:
            return {
                "success": False,
                "error": "Autoplac API key not configured"
            }
        
        # Example structure - actual API will differ
        payload = {
            "title": f"{car_data['brand']} {car_data['model']} {car_data['year']}",
            "description": car_data.get('description', ''),
            "price": car_data['price'],
            "mileage": car_data.get('mileage', 0),
            "year": car_data['year'],
            "fuel_type": car_data.get('fuel_type', ''),
            "transmission": car_data.get('transmission', ''),
            "images": car_data.get('images', []),
            "contact": {
                "phone": car_data.get('contact_phone', ''),
                "email": car_data.get('contact_email', '')
            }
        }
        
        try:
            # Placeholder for actual API call
            # response = requests.post(
            #     f"{self.api_url}/listings",
            #     json=payload,
            #     headers={"Authorization": f"Bearer {self.api_key}"}
            # )
            # return response.json()
            
            return {
                "success": True,
                "message": "Car published to Autoplac (placeholder)",
                "listing_id": "placeholder_id_12345"
            }
        
        except Exception as e:
            return {
                "success": False,
                "error": str(e)
            }
    
    def update_listing(self, listing_id, car_data):
        """
        Update an existing Autoplac listing
        
        Args:
            listing_id (str): Autoplac listing ID
            car_data (dict): Updated car information
            
        Returns:
            dict: Response from Autoplac API
        """
        # TODO: Implement actual update logic
        return {
            "success": True,
            "message": "Listing updated (placeholder)"
        }
    
    def delete_listing(self, listing_id):
        """
        Delete a listing from Autoplac
        
        Args:
            listing_id (str): Autoplac listing ID
            
        Returns:
            dict: Response from Autoplac API
        """
        # TODO: Implement actual delete logic
        return {
            "success": True,
            "message": "Listing deleted (placeholder)"
        }
